Move indexer into own folder for easier import
